refactor: Split file up into more lib files
