+++ crates/scheduler/src/cron.rs
@@ -1,0 +1,29 @@
+/*
+Copyright 2024-2025 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use std::{sync::Arc, time::Instant};
+
+use crate::ScheduleEvaluator;
+
+#[derive(Eq, PartialEq, Hash)]
+#[allow(dead_code)]
+pub struct CronSchedule(Arc<str>);
+impl ScheduleEvaluator for CronSchedule {
+    fn evaluate(&self) -> Instant {
+        // Implement the logic to evaluate the cron schedule
+        Instant::now()
+    }
+}
